get the certifiedTokens only once and stored as global variable
